M3: fix mypy no-any-return in _clock_s (CI Python Quality)
float()-cast the nanoseconds->seconds result so mypy sees a float return where rclpy is
untyped on the Layer-A runner (get_clock() -> Any with no rclpy installed). Mirrors _now_us's
int() cast. The earlier green local run was masked by a stale .mypy_cache; a clean rclpy-free
mypy now reproduces CI and passes.

+++ ros2_ws/src/patrol_mission/patrol_mission/node.py
@@ -172,7 +172,9 @@ class PatrolMissionNode(Node):
         )
 
     def _clock_s(self) -> float:
-        return self.get_clock().now().nanoseconds / 1e9
+        # float() so mypy sees a float return even where rclpy is untyped (nanoseconds -> Any on the
+        # Layer-A runner, which has no rclpy); mirrors _now_us's int() cast. Without it: no-any-return.
+        return float(self.get_clock().now().nanoseconds / 1e9)
 
     def _issue(self, cmd: Command) -> None:
         """Translate a decision-layer Command into /fmu/in/* messages.
